modified model/Sender.cc, cancel data map reference to Packet in Sender::findData, to reduce memory consumption

+++ model/Cache.h
@@ -189,8 +189,8 @@ private:
     //map filename  to index in files variable,because file is vector
     Name2index name2index;
     Files files;
-    uint32_t pkt_num = PKT_NUM;
-    uint32_t file_num = FILE_NUM;
+    uint8_t pkt_num = PKT_NUM;
+    uint8_t file_num = FILE_NUM;
     uint32_t cur_index = 0;
 
 public:
